Removed yaml templating changes and fixed ginkgo version2 errors for test run
