+++ go.mod
@@ -4,10 +4,9 @@ go 1.16
 
 require (
 	github.com/aws/aws-sdk-go v1.28.2
-	github.com/ghodss/yaml v1.0.0
 	github.com/go-logr/logr v0.4.0
 	github.com/google/uuid v1.1.2
-	github.com/onsi/ginkgo v1.16.5
+	github.com/onsi/ginkgo v1.16.4
 	github.com/onsi/ginkgo/v2 v2.1.1
 	github.com/onsi/gomega v1.17.0
 	github.com/openshift/api v0.0.0-20210805075156-d8fab4513288
