+++ tests/e2e/lib/kube_helpers.go
@@ -3,10 +3,8 @@ package lib
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"log"
 
-	"github.com/onsi/ginkgo"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -217,36 +215,5 @@ func isCredentialsSecretDeleted(namespace string, credSecretRef string) wait.Con
 		}
 		log.Printf("Secret still exists in namespace")
 		return false, err
-	}
-}
-
-func AreApplicationPodsRunning(namespace string) wait.ConditionFunc {
-	return func() (bool, error) {
-		clientset, err := setUpClient()
-		if err != nil {
-			return false, err
-		}
-		// select Velero pod with this label
-		veleroOptions := metav1.ListOptions{
-			LabelSelector: "e2e-app=true",
-		}
-		// get pods in test namespace with labelSelector
-		podList, err := clientset.CoreV1().Pods(namespace).List(context.TODO(), veleroOptions)
-		if err != nil {
-			return false, nil
-		}
-		if len(podList.Items) == 0 {
-			return false, nil
-		}
-		// get pod name and status with specified label selector
-		for _, podInfo := range podList.Items {
-			phase := podInfo.Status.Phase
-			if phase != corev1.PodRunning && phase != corev1.PodSucceeded {
-				ginkgo.GinkgoWriter.Write([]byte(fmt.Sprintf("Pod %v not yet succeeded", podInfo.Name)))
-				ginkgo.GinkgoWriter.Write([]byte(fmt.Sprintf("status: %v", podInfo.Status)))
-				return false, nil
-			}
-		}
-		return true, err
 	}
 }
